Fix LeetCode solution warnings

+++ Visual Studio/Misc/LeetCode OJ/015. 3Sum/Solution.h
@@ -21,7 +21,7 @@ public:
                 ++i;
                 while (i < num.size() && num[i] == value)
                 {
-                    v.emplace_back(i);
+                    v.emplace_back(static_cast<int>(i));
                     ++i;
                 }
             }
