+++ Visual Studio/Misc/LeetCode OJ/035. Search Insert Position/Solution.h
@@ -5,6 +5,6 @@ class Solution
 public:
     int searchInsert(int A[], int n, int target)
     {
-        return lower_bound(A, A + n, target) - A;
+        return static_cast<int>(lower_bound(A, A + n, target) - A);
     }
 };
